added "inputs" make target
this is handy -- it'll auto-download the inputs and store them in the inputs/
directory.  if the inputs are already there then it doens't download anything.

if we want to add another input, we just add a function to bin/inputs.py that
starts with "make_", and the rest is auto-magically taken care of.